+++ src/commands/chat/characters/character-update-subcommand.ts
@@ -6,32 +6,30 @@ import {
 	PermissionsString,
 	ComponentType,
 	ButtonStyle,
-	ButtonInteraction,
-	CacheType,
 } from 'discord.js';
+import { default as axios } from 'axios';
 
-import { EventData } from '../../../models/internal-models.js';
 import { InteractionUtils } from '../../../utils/index.js';
 import { Command, CommandDeferType } from '../../index.js';
 import { WgToken } from '../../../services/kobold/models/index.js';
 import { CharacterHelpers } from './helpers.js';
 import { CharacterUtils } from '../../../utils/character-utils.js';
-import { Language } from '../../../models/enum-helpers/index.js';
+import L from '../../../i18n/i18n-node.js';
 import { TranslationFunctions } from '../../../i18n/i18n-types.js';
 import { Config } from '../../../config/config.js';
 import { CharacterOptions } from './command-options.js';
-import { refs } from '../../../i18n/en/common.js';
+import { refs } from '../../../constants/common-text.js';
 import { PathBuilder } from '../../../services/pathbuilder/index.js';
 import { Creature } from '../../../utils/creature.js';
 import { CollectorUtils } from '../../../utils/collector-utils.js';
 import { KoboldError } from '../../../utils/KoboldError.js';
 
 export class CharacterUpdateSubCommand implements Command {
-	public names = [Language.LL.commands.character.update.name()];
+	public names = [L.en.commands.character.update.name()];
 	public metadata: RESTPostAPIChatInputApplicationCommandsJSONBody = {
 		type: ApplicationCommandType.ChatInput,
-		name: Language.LL.commands.character.update.name(),
-		description: Language.LL.commands.character.update.description(),
+		name: L.en.commands.character.update.name(),
+		description: L.en.commands.character.update.description(),
 		dm_permission: true,
 		default_member_permissions: undefined,
 	};
@@ -40,7 +38,6 @@ export class CharacterUpdateSubCommand implements Command {
 
 	public async execute(
 		intr: ChatInputCommandInteraction,
-		data: EventData,
 		LL: TranslationFunctions
 	): Promise<void> {
 		//check if we have an active character
@@ -85,11 +82,6 @@ export class CharacterUpdateSubCommand implements Command {
 				creature = Creature.fromPathBuilder(pathBuilderChar.build);
 			}
 
-			let result: {
-				intr: ButtonInteraction<CacheType>;
-				value: string;
-			};
-
 			if (creature.name !== activeCharacter.name) {
 				// confirm the update
 				const prompt = await intr.followUp({
@@ -121,7 +113,7 @@ export class CharacterUpdateSubCommand implements Command {
 					fetchReply: true,
 				});
 				let timedOut = false;
-				result = await CollectorUtils.collectByButton(
+				const result = await CollectorUtils.collectByButton(
 					prompt,
 					async buttonInteraction => {
 						if (buttonInteraction.user.id !== intr.user.id) {
@@ -149,7 +141,7 @@ export class CharacterUpdateSubCommand implements Command {
 						},
 					}
 				);
-				if (result.value !== 'update') {
+				if (result && result.value !== 'update') {
 					await InteractionUtils.editReply(intr, {
 						content: LL.sharedInteractions.choiceRegistered({
 							choice: 'Cancel',
@@ -235,7 +227,7 @@ export class CharacterUpdateSubCommand implements Command {
 				);
 			} catch (err) {
 				console.log(err);
-				if (err?.response?.status === 401) {
+				if ((axios.default ?? axios).isAxiosError(err) && err?.response?.status === 401) {
 					//token expired!
 					await WgToken.query().delete().where({ charId: activeCharacter.charId });
 					await InteractionUtils.send(
@@ -251,7 +243,10 @@ export class CharacterUpdateSubCommand implements Command {
 						true
 					);
 					return;
-				} else if (err?.response?.status === 429) {
+				} else if (
+					(axios.default ?? axios).isAxiosError(err) &&
+					err?.response?.status === 429
+				) {
 					await InteractionUtils.send(
 						intr,
 						LL.commands.character.interactions.tooManyWGRequests()
@@ -259,7 +254,11 @@ export class CharacterUpdateSubCommand implements Command {
 					return;
 				} else {
 					//otherwise, something else went wrong that we want to be a real error
-					throw err;
+
+					console.error(err);
+					throw new KoboldError(
+						`Yip! Something went wrong when I tried to fetch the character update from wanderer's guide!.`
+					);
 				}
 			}
